refactor(user): refactor user to support DI
