+++ server/api/user/userController.test.js
@@ -1,7 +1,9 @@
 /* eslint no-underscore-dangle: ["error", { "allow": ["_id", "_doc"] }] */
 const mockingoose = require('mockingoose').default;
-const userController = require('./userController');
+const UserController = require('./userController');
 
+let userController;
+let middleware;
 let req = null;
 let res = null;
 let next = null;
@@ -9,9 +11,15 @@ let next = null;
 describe('User Controller API', () => {
   beforeEach(() => {
     mockingoose.resetAll();
+    middleware = {
+      appValidation: {
+        validateRequest: jest.fn(),
+      },
+    };
     req = jest.fn();
     res = jest.fn();
     next = jest.fn();
+    userController = UserController({ middleware });
   });
 
   it('Should return an error APIError with status 403 and Invalid id when an CastError occurs', async () => {
@@ -77,6 +85,7 @@ describe('User Controller API', () => {
   it('Should return an error when user.save throws an generic error', async () => {
     const genericError = new Error('Any error!');
     mockingoose.users.toReturn(genericError, 'save');
+
     req = {
       body: { username: 'myUsername', password: 'myPassword', role: 'myRole' },
     };
