+++ server/api/user/userRoutes.js
@@ -1,8 +1,10 @@
 const router = require('express').Router();
-const userController = require('./userController');
+const UserController = require('./userController');
 const userValidator = require('./userValidator');
 
-const userRoutes = () => {
+const userRoutes = ({ middleware }) => {
+  const userController = UserController({ middleware });
+
   router.param('id', userController.params);
   router.get('/me', userController.me);
