+++ server/middleware/index.js
@@ -1,8 +1,10 @@
 const appMiddleware = require('./appMiddleware');
 const errorHandlingMiddleware = require('./errorHandlingMiddleware');
+const validation = require('./validation');
 
 const middleware = ({ config, logger }) => ({
   appMiddleware: appMiddleware({ config }),
+  appValidation: validation,
   errorHandlingMiddleware: errorHandlingMiddleware({ logger }),
 });
